Add CAL_SIM_POST_LAUNCH_WAIT env var
Controls how long to wait _after_ the simulator is opened.

In CoreSimulator environments, the iOS Simulator starts many
async processes that must be allowed to finish before we start
operating on the simulator.  Until we find the right combination
of processes to wait for, this variable will give us the
opportunity to control how long we wait.

Essential for managed envs like Travis + Jenkins and on slower
machines.