CoreSim#installed_app_bundle_dir stop using 'simctl'
simctl get_app_container requires the device to be
booted on El Cap.  On Yosemite, it works without the
device being booted.  Testing shows that moving the
device in and out of the booted state is problematic.

To address incomplete app installations (the reason for
changing over to simctl), the #installed_app_bundle_dir
checks the install and cleans up incomplete installations.